Add support for `SELECT ... FOR UPDATE`
The semantics of where this modifier is allowed are taken from PG, even
though this query is also allowed on MySQL. PG's docs are much more
thorough on the semantics of `FOR UPDATE`, and where it is or is not
allowed.

I've opted *not* to allow this with boxed queries, as a boxed query has
no way to verify that it isn't mixed with a group by/having/distinct
clause. We may want to lift this restriction in the future, depending
on what we do with `group_by` there when we add proper support for it.
If we decide to allow `group_by` on boxed queries, it will require not
that the select clause is still valid. It seems likely that boxed
queries will just end up with the gotcha of "less checking is done".

Still, for the time being I've taken the conservative approach, as we
can always change that later.

If/when we add support for having, union, intersect, and except, we will
need to ensure that those methods can only be used with no for update
clause as well.

The code is structured in a way that's a bit more generic on the backend
than is strictly necessary, but I plan to follow this up with support
for MySQL's `LOCK IN SHARE MODE` (which I believe is equivalent to `FOR
SHARE` in PG), which will need this structure in place.

I've opted for a very minimal usage example, as the test which
demonstrates its behavior is long enough that it drowns out the actual
usage.