Implement `Insertable` for slice and vec rather than any iterator
This blanket impl conflicts with an impl for tuples (since tuples and
`IntoIterator` are non-local, we are not allowed to rely on the fact
that `&(A, B): !IntoIterator`). This impl would have only been useable
with tuples anyway, since `IntoInsertStatement` is implemented directly
on vec and slice, not arbitrary iterators.

Since the `Insertable` impl could now rely on slice specifically, there
was no reason for me to keep the nonsense we were doing for
`BatchInsertValues` remaining generic, either.