rm `BatchInsertStatement`
This brings us back to the place we want to be, where there is only one
`InsertStatement` struct. Rather than executing 0 queries when inserting
an empty slice, we instead execute a select statement that will return 0
rows (effectively turning it into a no-op query).

It'd be technically more correct for us to stick the returning clause in
the select statement, so any assertions about the number of columns or
their types would still pass. However, that implementation would be a
bit harder to write, so for now I've opted to rely on the fact that we
will never make those assertions when we have no rows to return.

With this change, we can use the blanket impl of `ExecuteDsl` for the
"normal" case (though I had to modify it a tad to ensure it is disjoint
with the sqlite slice impl, since associated types aren't taken into
account when determining disjointness).

One alternative implementation would be to remove the blanket impl
entirely. This might be something we want to do in the future, since
right now `foo.eq(bar).execute(&conn)` actually compiles. However, we'd
need to remove the blanket impl of `LoadDsl` as well if we wanted to
elminate this select statement, which I'm less inclined to do.