Allow `filter` to be called on `UpdateStatement` and `DeleteStatement` (#1192)
This change is in the same spirit as #1166. By allowing the method to be
chained as part of the query, instead of on the argument to a function,
it becomes much easier to introduce line breaks naturally.

Why isn't this `impl FilterDsl`?
--------------------------------

Today `FilterDsl` is structured in such a way that it assumes that it's
only implemented on valid complete queries, and that it will return a
valid complete query (partially to ensure that the call to `filter`
can't change the SQL type). Since these structs are only complete
queries when they have a returning clause (and
`IncompleteUpdateStatement` is never a complete query), we would have to
loosen those bounds.

We may want to make that change eventually, but for now I opted to
implement these as inherent methods.

Why not add `find` as well?
---------------------------

We could. To me, `update(users).set(...).find(1)` just feels off. Given
that `find` is typically much shorter, I've opted to exclude it for now.