Shorten the changelog entry for #1231 [ci skip]
This entry was very long, and didn't make it clear what action was
required by our users. I've shortened it to the 4 things that actually
affect users of Diesel. (This was an awesome PR, but the internal
changes will be seen by contributors without being listed in the user
facing changelog)