Document the SQLite module
I'm going to start adding `#[deny(missing_docs)]` as I go through these
now. There's not a ton to say about any of these types, and some of
these doc comments are part of why I generally hate
`#[deny(missing_docs)]`, but I still think it will be a net win to have
that on Diesel.