Fix compile-fail tests
Unfortunately, several of our errors actually got much worse. We have a
bunch of overflows where we didn't before (I am convinced this is a bug
in rustc), and attempting to box an invalid query now just complains
about `BoxedDsl` not being implemented, rather than the reason that
`BoxedDsl` is not implemented. I'm not sure what the reason is for the
second one.