Rewrite the `select` docs
While most of the bounds related here are really more relevant to those
traits than `select`, but this is where users will encounter it, so I
think it's worth mentioning here. I'm unsure if we should mention the
aggregate/non-aggregate stuff here or not, but I've left it off for now.