Use `diesel print-schema` in our tests
`infer_schema!` is meant to be a useful tool for when you're just
getting started on a project. Most projects should eventually migrate to
using `diesel print-schema`, once their schema is being modified less
often.

Even though the hypothetical project in the guide would be using
`infer_schema!` at this point in time, I think that for teaching
purposes, we should use show the generated code.