Require `Debug` on `Connection::test_transaction` errors
I noticed while writing the insert guide that if something inside of
`test_transaction` returns an error, the panic just says "Transaction
did not succeed" but doesn't say why.

I don't want to deal with fixing this right now, but ultimately to do so
we will need either `Display`, `Debug`, or `Error` as a bound, which is
a semver breaking change.

I've opted to use `Debug` as the bound so that we can switch to use
`Result::expect` later on.