Guide draft: Composing Applications with Diesel
This guide goes into how to write functions which return both queries
and expressions, how to use the helper types, and how to use their boxed
forms.

This was originally going to be part of the advanced querying guide, but
that guide got too large so I split out an "extending diesel" guide from
it, but this felt separate from everything else in that guide, so I
decided it made sense as its own stand-alone guide

Unlike our other guides, there's no code in the examples directory for
this one. The links are going to point at crates.io. The code is
slightly modified from the real thing, but I'd still prefer to point the
links at the real app than our example copy of it.